requirements.txt
Django>=2.2.7
django-rest-framework>=0.1.0
psycopg2>=2.8.4
freezegun>=0.3.12
